Moved search highlight code to SolrFacetService.
